api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0